+++ api/test/core/timestamp_test.cc
@@ -1,9 +1,9 @@
-#include "opentelemetry/core/timestamp.h"
+#include "opentelemetry/common/timestamp.h"
 
 #include <gtest/gtest.h>
 
-using opentelemetry::core::SteadyTimestamp;
-using opentelemetry::core::SystemTimestamp;
+using opentelemetry::common::SteadyTimestamp;
+using opentelemetry::common::SystemTimestamp;
 
 template <class Timestamp>
 static bool AreNearlyEqual(const Timestamp &t1, const Timestamp &t2) noexcept