+++ sdk/src/trace/span.cc
@@ -11,8 +11,8 @@ namespace sdk
 namespace trace
 {
 
-using opentelemetry::core::SteadyTimestamp;
-using opentelemetry::core::SystemTimestamp;
+using opentelemetry::common::SteadyTimestamp;
+using opentelemetry::common::SystemTimestamp;
 
 namespace
 {
@@ -129,7 +129,7 @@ void Span::AddEvent(nostd::string_view name) noexcept
   recordable_->AddEvent(name);
 }
 
-void Span::AddEvent(nostd::string_view name, core::SystemTimestamp timestamp) noexcept
+void Span::AddEvent(nostd::string_view name, SystemTimestamp timestamp) noexcept
 {
   std::lock_guard<std::mutex> lock_guard{mu_};
   if (recordable_ == nullptr)
@@ -140,7 +140,7 @@ void Span::AddEvent(nostd::string_view name, core::SystemTimestamp timestamp) no
 }
 
 void Span::AddEvent(nostd::string_view name,
-                    core::SystemTimestamp timestamp,
+                    SystemTimestamp timestamp,
                     const opentelemetry::common::KeyValueIterable &attributes) noexcept
 {
   std::lock_guard<std::mutex> lock_guard{mu_};